Move runtime_types as well
