+++ impeller/runtime_stage/runtime_stage.h
@@ -10,7 +10,7 @@
 #include "flutter/fml/macros.h"
 #include "flutter/fml/mapping.h"
 
-#include "flutter/impeller/runtime_stage/runtime_types.h"
+#include "flutter/impeller/core/runtime_types.h"
 
 namespace impeller {
